sagemaker/code/requirements.txt
numpy
rdkit-pypi
torch